refactor: use single startFirebaseUI, move firebaseAdmin, and remove shared-firebase
