+++ boilerplates/firebase-auth/files/libs/firebaseAdmin.ts
