+++ boilerplates/h3/package.json
@@ -14,7 +14,7 @@
   "devDependencies": {
     "@auth/core": "^0.28.0",
     "@batijs/compile": "workspace:*",
-    "@batijs/shared-firebase": "workspace:*",
+    "@batijs/firebase-auth": "workspace:*",
     "@batijs/trpc": "workspace:^",
     "@hattip/polyfills": "^0.0.43",
     "@trpc/server": "^10.45.1",
