+++ boilerplates/react-firebase-auth/files/pages/login/+Page.tsx
@@ -1,6 +1,6 @@
 import "firebaseui/dist/firebaseui.css"
 import React, { useEffect, useState } from 'react'
-import firebase from "firebase/compat/app"
+import { startFirebaseUI } from "@batijs/firebase-auth/libs/firebaseUI"
 import * as firebaseui from "firebaseui"
 import { getAuth, type UserCredential } from 'firebase/auth'
 import { reload } from "vike/client/router"
@@ -10,41 +10,31 @@ export default Page
 function Page() {
     const [error, setError] = useState("")
 
+    async function sessionLogin(authResult: UserCredential) {
+        const idToken = await authResult.user.getIdToken() || ""
+        try {
+            const response = await fetch("/api/sessionLogin", {
+                method: "POST",
+                body: JSON.stringify({ idToken }),
+                headers: {
+                    "Content-Type": "application/json",
+                },
+            })
+            if (response.ok) {
+                await reload()
+            } else {
+                setError(response.statusText)
+            }
+            await getAuth().signOut()
+        } catch (err) {
+            console.log('error :', err)
+        }
+    }
+    
     useEffect(() => {
         const ui = firebaseui.auth.AuthUI.getInstance() || new firebaseui.auth.AuthUI(getAuth())
         if (!error) {
-            // Please read firebaseui docs at https://github.com/firebase/firebaseui-web
-            ui.start("#firebaseui-auth-container", {
-                callbacks: {
-                    signInSuccessWithAuthResult(authResult: UserCredential) {
-                        authResult.user.getIdToken().then((idToken) => {
-                            fetch("/api/sessionLogin", {
-                                method: "POST",
-                                body: JSON.stringify({ idToken }),
-                                headers: {
-                                    "Content-Type": "application/json",
-                                },
-                            }).then(async (response) => {
-                                if (response.ok) {
-                                    await reload()
-                                } else {
-                                    setError(response.statusText)
-                                }
-                                await getAuth().signOut()
-                            })
-                        })
-                        // Don't redirect after firebase client successfully sign-in, let vike handle the rest.
-                        return false
-                    }
-                },
-                signInFlow: "popup",
-                signInOptions: [
-                    // Disable "Email enumeration protection" to be able to login with registered email address
-                    // https://console.firebase.google.com/u/1/project/{project-id}/authentication/settings
-                    firebase.auth.EmailAuthProvider.PROVIDER_ID,
-                    firebase.auth.GoogleAuthProvider.PROVIDER_ID,
-                ],
-            })
+            startFirebaseUI(ui, sessionLogin)
         }
     }, [error])
