+++ boilerplates/react-firebase-auth/package.json
@@ -13,6 +13,7 @@
   "license": "MIT",
   "devDependencies": {
     "@batijs/compile": "workspace:*",
+    "@batijs/firebase-auth": "workspace:*",
     "@types/node": "^18.19.14",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
