--- boilerplates/shared-firebase/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@batijs/shared-firebase",
-  "private": true,
-  "version": "0.0.1",
-  "description": "",
-  "type": "module",
-  "scripts": {
-    "check-types": "tsc --noEmit",
-    "build": "bati-compile-boilerplate"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "@batijs/compile": "workspace:*",
-    "@types/node": "^18.19.14",
-    "dotenv": "^16.4.4",
-    "firebase-admin": "^12.0.0"
-  },
-  "dependencies": {
-    "@batijs/core": "workspace:*"
-  },
-  "files": [
-    "dist/"
-  ],
-  "bati": {
-    "if": {
-      "flag": {
-        "$in": [
-          "firebase-auth"
-        ]
-      }
-    }
-  },
-  "exports": {
-    "./firebaseAdmin": {
-      "types": "./dist/types/firebaseAdmin.d.ts"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "firebaseAdmin": [
-        "./dist/types/firebaseAdmin.d.ts"
-      ]
-    }
-  }
-}
