+++ boilerplates/solid-firebase-auth/package.json
@@ -13,6 +13,7 @@
   "license": "MIT",
   "devDependencies": {
     "@batijs/compile": "workspace:^",
+    "@batijs/firebase-auth": "workspace:*",
     "@types/node": "^18.19.14",
     "firebase": "^10.8.0",
     "firebaseui": "^6.1.0",
