+++ boilerplates/vue-firebase-auth/files/pages/login/+Page.vue
@@ -9,53 +9,43 @@
 <script setup lang="ts">
 import "firebaseui/dist/firebaseui.css"
 import { ref, onMounted, onUpdated } from "vue";
-import firebase from "firebase/compat/app"
 import * as firebaseui from "firebaseui"
 import { getAuth, type UserCredential } from 'firebase/auth'
 import { reload } from "vike/client/router"
+import { startFirebaseUI } from "@batijs/firebase-auth/libs/firebaseUI"
+
 const error = ref("")
 const clearError = () => error.value = ""
 const ui = firebaseui.auth.AuthUI.getInstance() || new firebaseui.auth.AuthUI(getAuth())
 
-const uiCOnfig = {
-    callbacks: {
-        signInSuccessWithAuthResult(authResult: UserCredential) {
-            authResult.user.getIdToken().then((idToken) => {
-                fetch("/api/sessionLogin", {
-                    method: "POST",
-                    body: JSON.stringify({ idToken }),
-                    headers: {
-                        "Content-Type": "application/json",
-                    },
-                }).then(async (response) => {
-                    if (response.ok) {
-                        await reload()
-                    } else {
-                        error.value = response.statusText
-                    }
-                    await getAuth().signOut()
-                })
-            })
-            // Don't redirect after firebase client successfully sign-in, let vike handle the rest.
-            return false
+async function sessionLogin(authResult: UserCredential) {
+    const idToken = await authResult.user.getIdToken() || ""
+    try {
+        const response = await fetch("/api/sessionLogin", {
+            method: "POST",
+            body: JSON.stringify({ idToken }),
+            headers: {
+                "Content-Type": "application/json",
+            },
+        })
+        if (response.ok) {
+            await reload()
+        } else {
+            error.value = response.statusText
         }
-    },
-    signInFlow: "popup",
-    signInOptions: [
-        // Disable "Email enumeration protection" to be able to login with registered email address
-        // https://console.firebase.google.com/u/1/project/{project-id}/authentication/settings
-        firebase.auth.EmailAuthProvider.PROVIDER_ID,
-        firebase.auth.GoogleAuthProvider.PROVIDER_ID,
-    ],
+        await getAuth().signOut()
+    } catch (err) {
+        console.log('error :', err)
+    }
 }
 
 onMounted(() => {
-    ui.start("#firebaseui-auth-container", uiCOnfig)
+    startFirebaseUI(ui, sessionLogin)
 })
 
 onUpdated(() => {
     if (!error.value) {
-        ui.start("#firebaseui-auth-container", uiCOnfig)
+        startFirebaseUI(ui, sessionLogin)
     }
 })
 </script>
